Cleaning of c_syntax, comp_sections, misc, movements, prettyprint, ri-util. Move some functions to misc and ri-util. Introduce a new get_variable_name_t to ease variable_name method passing without warning. Changed to compile on Linux x86_64, but still compile on x86_32 and validate. Remove many compilation warnings in both versions. Some presentation cleaning.
git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@12833 32a0c1e6-9308-0410-80dd-f9d2257a62c7

+++ src/Libs/comp_sections/ss.c
@@ -449,7 +449,7 @@ AllocateDadStruct(int Rank)
 {
   dad_struct Dad;
   bound_pair Bounds;
-  unsigned cnt;
+  int cnt;
   list RefTempList = NIL;
   list BoundList = NIL;
   rtype Ref = make_rtype(is_rtype_nonlinear, UU);
@@ -621,7 +621,7 @@ TransRefTemp(loop ThisLoop, comp_desc Desc)
 {
   tVariants Vars;
   simple_section Dad;
-  unsigned DimNo, Rank;
+  intptr_t DimNo, Rank;
   normalized Nexpr;
   expression TmpExpr;
   list OldList, NewList, NewEle;
